==> notebooks/04b-similarity_scores_duration.ipynb <==
# change root path one level up
import os
os.chdir("..")

import numpy as np
import pandas as pd
import pytz
import pickle
import time

# lightgbm and optuna
from sklearn.metrics.pairwise import cosine_similarity
from tqdm.notebook import tqdm
# import datetime and timedelta
from datetime import datetime, timedelta




# CAT_FEATURES = ['stationID', 'spaceID', 'siteID', 'clusterID', 'userID', 'paymentRequired', 'connectionTime_Weekday', 'connectionTime_Hour', 'connectionTime_Month', 'connectionTime_is_holiday', 'connectionTime_is_weekend', 'connectionTime_weekday_hour']
CAT_FEATURES = ["userID"]
NUM_FEATURES = [
    "connectionTime_hour_x",
    "connectionTime_hour_y",
    "connectionTime_month_x",
    "connectionTime_month_y",
    "connectionTime_weekday_x",
    "connectionTime_weekday_y",
]
TARGET = 'parking_time'


sessions = pd.read_csv("./data/caltech_test_data.csv", index_col=[0, 1])
for cols in CAT_FEATURES:
    sessions[cols] = sessions[cols].astype("category")
sessions = sessions[NUM_FEATURES + CAT_FEATURES + [TARGET]]
sessions


# find dates 30, 60, 120, 240, 360, 480, 600 days before 2020-12-01 and make a list of them
dates = []
for i in [30, 60, 120, 240, 360, 480]:
    dates.append((datetime(2019, 11, 1) - timedelta(days=i)).strftime("%Y-%m-%d"))


# make 6 partitions for training starting with dates[i] and ending with datetime(2020, 12, 1)
train_sessions = []
for i in range(len(dates)):
    train_sessions.append(sessions.loc[dates[i] : "2019-12-01"])
    # print number of sessions in each partition
    print(train_sessions[i].shape)


# testing data is collected from Dec. 1, 2018 to Jan. 1, 2019
test = sessions.loc[(slice("2019-12-01", "2020-01-01"), slice(None)), :]

def get_similar_sessions(train_test_ds, feats, session_idx, n=5, target=TARGET):
    train_ds_scaled = train_test_ds[feats]
    # start timer
    start = time.time()
    # if categorical features are in feats
    if any([x in CAT_FEATURES for x in feats]):
        # dummy encode categorical features
        train_ds_scaled = pd.get_dummies(train_ds_scaled, columns=CAT_FEATURES)
    historical = train_ds_scaled[:session_idx]
    # calculate the cosine similarity for the row of session_idx
    cosine_sim = cosine_similarity(
        historical, train_ds_scaled.loc[session_idx].values.reshape(1, -1)
    )
    # make a dataframe
    cosine_sim_df = pd.DataFrame(cosine_sim, index=historical.index)
    # show the top 10 most similar sessions
    top = (
        train_test_ds[:session_idx]
        .loc[cosine_sim_df.nlargest(n, columns=0).index][target]
        .values
    )
    # calculate mean of top n similar sessions
    top_mean = top.mean()
    runtime = time.time() - start
    # clean up
    del train_ds_scaled, historical, cosine_sim, cosine_sim_df, top
    return top_mean, runtime

# evaluate smape of lightgbm model
def smape(y_true, y_pred):
    return np.mean(np.abs((y_true - y_pred) / (y_true + y_pred))) * 100


def mae(y_true, y_pred):
    return np.mean(np.abs(y_true - y_pred))

# save distribution of errors
def save_error_dist(y_true, y_pred, name):
    error = y_pred - y_true
    # save as a picle file
    with open('./results/'+name+'.pkl', "wb") as f:
        pickle.dump(error, f)

results = pd.DataFrame(
    columns=["smape", "mae"], index=[30, 60, 120, 240, 360, 480]
)
for i, ds in enumerate(train_sessions):
    # size of the dataset
    print(ds.shape)
    concat = pd.concat(
        [ds[NUM_FEATURES + [TARGET]], test[NUM_FEATURES + [TARGET]]]
    )
    test_fore = test.copy()
    rt_list = []
    for j in range(20):
        for idx in test.index.tolist():
            # get similar sessions
            top_mean, rt = get_similar_sessions(concat, NUM_FEATURES, idx)
            test_fore.loc[idx, TARGET+"_forecast"] = top_mean
            rt_list.append(rt)
    print(f"Runtime: {np.mean(rt_list)}")
    # evaluate smape of similar sessions model
    print(
        "smape of similar sessions model: ",
        smape(test_fore[TARGET], test_fore[TARGET+"_forecast"]),
    )
    results.iloc[i]["smape"] = smape(
        test_fore[TARGET], test_fore[TARGET+"_forecast"]
    )
    # evaluate mae of similar sessions model
    print(
        "mae of similar sessions model: ",
        mae(test_fore[TARGET], test_fore[TARGET+"_forecast"]),
    )
    results.iloc[i]["mae"] = mae(
        test_fore[TARGET], test_fore[TARGET+"_forecast"]
    )
    dist = save_error_dist(test_fore[TARGET], test_fore[TARGET+"_forecast"], "similar_sessions_duration_"+str(results.index[i]))
    


# save to csv
results.to_csv("./results/similar_sessions_duration_nologs.csv")
results


results = pd.DataFrame(
    columns=["smape", "mae"], index=[30, 60, 120, 240, 360, 480]
)
for i, ds in enumerate(train_sessions):
    # size of the dataset
    print(ds.shape)
    concat = pd.concat(
        [ds[NUM_FEATURES + CAT_FEATURES +[TARGET]], test[NUM_FEATURES + CAT_FEATURES + [TARGET]]]
    )
    test_fore = test.copy()
    rt_list = []
    for j in range(20):
        for idx in test.index.tolist():
            # get similar sessions
            top_mean, rt = get_similar_sessions(concat, NUM_FEATURES + CAT_FEATURES, idx)
            test_fore.loc[idx, TARGET+"_forecast"] = top_mean
            rt_list.append(rt)
    print(f"Runtime: {np.mean(rt_list)}")
    # evaluate smape of similar sessions model
    print(
        "smape of similar sessions model: ",
        smape(test_fore[TARGET], test_fore[TARGET+"_forecast"]),
    )
    results.iloc[i]["smape"] = smape(
        test_fore[TARGET], test_fore[TARGET+"_forecast"]
    )
    # evaluate mae of similar sessions model
    print(
        "mae of similar sessions model: ",
        mae(test_fore[TARGET], test_fore[TARGET+"_forecast"]),
    )
    results.iloc[i]["mae"] = mae(
        test_fore[TARGET], test_fore[TARGET+"_forecast"]
    )
    save_error_dist(test_fore[TARGET], test_fore[TARGET+"_forecast"], "similar_sessions_duration_"+str(results.index[i]))
    


# save to csv
results.to_csv("./results/similar_sessions_duration.csv")
results
