==> notebooks/05-user_input.ipynb <==
# change root path one level up
import os
os.chdir("../acnportal/")


# lightgbm and optuna
import lightgbm as lgb
import matplotlib.pyplot as plt
import numpy as np
import optuna
import pandas as pd
import pytz
import seaborn as sns
from tqdm.notebook import tqdm

CAT_FEATURES = [
    "stationID",
    "spaceID",
    "siteID",
    "clusterID",
    "userID",
    "paymentRequired",
    "connectionTime_Weekday",
    "connectionTime_Hour",
    "connectionTime_Month",
    "connectionTime_is_holiday",
    "connectionTime_is_weekend",
    "connectionTime_weekday_hour",
]


sessions = pd.read_csv("../data/caltech_test_data.csv", index_col=[0, 1])
# drop idle_time column
sessions = sessions.drop(columns=["idle_time"])
sessions


# testing data is collected from Dec. 1, 2018 to Jan. 1, 2019
test = sessions.loc[(slice("2018-12-01", "2019-01-01"), slice(None)), :]


# evaluate smape
def smape(y_true, y_pred):
    return np.mean(np.abs(y_true - y_pred) / np.abs(y_true + y_pred)) * 100


# evaluate smape of requested parking time
print(
    "smape of requested parking time: ",
    smape(test["Requested_parking_time"], test["parking_time"]),
)


# evaluare mae of requested parking time
print(
    "mae of requested parking time: ",
    np.mean(np.abs(test["Requested_parking_time"] - test["parking_time"])),
)


# evaluate smape of requested energy
print("smape of requested energy: ", smape(test["kWhRequested"], test["kWhDelivered"]))


# evaluare mae of requested energy
print(
    "mae of requested energy: ",
    np.mean(np.abs(test["kWhRequested"] - test["kWhDelivered"])),
)


# put all metrics in a dataframe
metrics = pd.DataFrame(
    {
        "smape of requested parking time": [
            smape(test["Requested_parking_time"], test["parking_time"])
        ],
        "mae of requested parking time": [
            np.mean(np.abs(test["Requested_parking_time"] - test["parking_time"]))
        ],
        "smape of requested energy": [
            smape(test["kWhRequested"], test["kWhDelivered"])
        ],
        "mae of requested energy": [
            np.mean(np.abs(test["kWhRequested"] - test["kWhDelivered"]))
        ],
    }
)
# save metrics to csv
metrics.to_csv("./metrics_user_input.csv")
metrics
